[#93] - move KakaoCallback method from RegistrationController to LoginController
