+++ src/main/java/io/github/mygoodsupporter/controller/MemberController.java
@@ -1,15 +1,8 @@
 package io.github.mygoodsupporter.controller;
 
 
-import io.github.mygoodsupporter.security.MemberDetails;
 import lombok.RequiredArgsConstructor;
-import org.aspectj.bridge.MessageUtil;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.*;
-import io.github.mygoodsupporter.domain.Member;
-import io.github.mygoodsupporter.service.MemberService;
-
 
 
 @Controller
