+++ src/main/java/io/github/mygoodsupporter/domain/kakao/OAuthToken.java
@@ -1,4 +1,4 @@
-package io.github.mygoodsupporter.domain;
+package io.github.mygoodsupporter.domain.kakao;
 
 
 import lombok.Getter;
