+++ src/main/java/io/github/mygoodsupporter/domain/member/Member.java
@@ -1,7 +1,8 @@
-package io.github.mygoodsupporter.domain;
+package io.github.mygoodsupporter.domain.member;
 
-import lombok.*;
-
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 import java.util.ArrayList;
 import java.util.List;
