+++ src/main/java/io/github/mygoodsupporter/security/MemberDetailsService.java
@@ -1,7 +1,7 @@
 package io.github.mygoodsupporter.security;
 
 import io.github.mygoodsupporter.dao.MemberDAO;
-import io.github.mygoodsupporter.domain.Member;
+import io.github.mygoodsupporter.domain.member.Member;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
