+++ src/test/java/io/github/mygoodsupporter/dao/MemberDAOTest.java
@@ -1,6 +1,6 @@
 package io.github.mygoodsupporter.dao;
 
-import io.github.mygoodsupporter.domain.Member;
+import io.github.mygoodsupporter.domain.member.Member;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
